+++ cmd/mount.go
@@ -216,7 +216,7 @@ var mountCmd = &cobra.Command{
 		if !disableVersionCheck {
 			err := VersionCheck()
 			if err != nil {
-				return err
+				fmt.Fprintf(os.Stderr, "%s", err.Error())
 			}
 		}