+++ cmd/mountv1.go
@@ -153,7 +153,7 @@ var generateConfigCmd = &cobra.Command{
 		if !disableVersionCheck {
 			err := VersionCheck()
 			if err != nil {
-				return err
+				fmt.Fprintf(os.Stderr, "%s", err.Error())
 			}
 		}
 		resetOptions()